fixing builds and moving test out
