+++ setup.py
@@ -1,7 +1,9 @@
 from setuptools import setup
 
-setup(name='distributions',
-      version='0.1',
-      description='Gaussian distributions',
-      packages=['distributions'],
+setup(name='statistics_prob_distributions',
+      version='1',
+      description='Gaussian distributions and Binomial',
+      packages=['statistics_prob_distributions'],
+      author='Muhammad Rizwan',
+      author_email='xrizwanr@gmail.com',
       zip_safe=False)
