+++ statistics_prob_distributions/Binomialdistribution.py
@@ -1,6 +1,6 @@
-from distributions import Distribution
 import math
 import matplotlib.pyplot as plt
+from .Generaldistribution import Distribution
 
 class Binomial(Distribution):
     """ Binomial distribution class for calculating and
