+++ statistics_prob_distributions/Generaldistribution.py
