+++ statistics_prob_distributions/README.md
@@ -1,0 +1,1 @@
+# Statistics probabilty package
